Added main_menu package. Restructuring.
