--- app/src/main/java/eu/ezytarget/micopi/LaunchSelectionListener.kt
@@ -1,6 +1,0 @@
-package eu.ezytarget.micopi
-
-interface LaunchSelectionListener {
-
-    fun onContactPickerSelected(allowMultipleSelection: Boolean = false)
-}
