+++ app/src/main/java/eu/ezytarget/micopi/main_menu/ContactPickerResultHandler.kt
@@ -1,0 +1,4 @@
+package eu.ezytarget.micopi.main_menu
+
+class ContactPickerResultHandler {
+}
