+++ app/src/main/java/eu/ezytarget/micopi/main_menu/MainMenuSelectionListener.kt
@@ -1,0 +1,7 @@
+package eu.ezytarget.micopi.main_menu
+
+interface MainMenuSelectionListener {
+
+    fun onContactPickerSelected(allowMultipleSelection: Boolean = false)
+    
+}
