+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/launchBackgroundColor"
-    tools:context=".MainActivity">
+    tools:context=".main_menu.MainActivity">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
